fx.Module: Reorganize code
In #830, I reverted some of the code moves to make reviewing the change
easier. This change adds back those moves on top of that PR.

This reverts commit e4d006b9cb4eb6eede8e1ca672cb4d74f679ddf2.
